backend/requirements.txt
fastapi>=0.115.6
haystack-ai==2.8.0
markdown-it-py>=3.0.0
mdit_plain>=1.0.1
opensearch-haystack==1.2.0
pydantic-settings>=2.7.0
pypdf>=5.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.19
sentence-transformers>=3.3.1
uvicorn>=0.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
sqlalchemy>=2.0.0
httpx>=0.25.0